+++ pages/info/service/serviceData.js
@@ -5,7 +5,7 @@ export const servicesData = [
         "Nemo enim ipsam voluptatem quia voluptas sit aspernatur aut odit aut fugit, sed quia consequuntur magni dolores",
       image: "/pharmacy.jpg",
       alt: "tablets",
-      nextPage: "/info/pharmacy"
+      nextPage: "/workwithus/pharmacy"
     },
     {
       title: "Laboratory",
@@ -13,7 +13,7 @@ export const servicesData = [
         "Eos qui ratione voluptatem sequi nesciunt. Neque porro quisquam est, qui dolorem ipsum quia dolor sit amet, consectetur, adipisci.",
       image: "/laboratory.jpg",
       alt: "pharmacist looking at the microscope",
-      nextPage: "/info/laboratory"
+      nextPage: "/workwithus/laboratory"
     },
     {
       title: "Consultation",
@@ -21,6 +21,6 @@ export const servicesData = [
         "Quis autem vel eum iure reprehenderit qui in ea voluptate velit esse quam nihil molestiae consequatur, vel illum qui dolorem eum fugiat quo voluptas nulla pariatur",
       image: "/consultation.jpg",
       alt: "patient talking to the doctor through videocall",
-      nextPage: "/info/consultation"
+      nextPage: "/workwithus/consultation"
     },
   ];