+++ pages/info/terms.js
@@ -4,74 +4,29 @@ import SectionTitle from '../../components/SectionTitle';
 
 import styles from '../../styles/Terms.module.css'
 
+import { termsAndConditions } from './service/termsAndConditions';
+
 export default function terms() {
-  return (
-    <section class="section is-medium has-text-centered">
-      <SectionTitle name="Terms and Conditions" />
-      <div class="container box is-size-5-mobile">
-        <div className={styles.terms}>
-          <ol>
-            <li>
-              <strong>Lorem ipsum dolor sit amet</strong> <br /> consectetur
-              adipiscing elit, sed do eiusmod tempor incididunt ut labore et
-              dolore magna aliqua. Ut enim ad minim veniam, quis nostrud
-              exercitation ullamco laboris nisi ut aliquip ex ea commodo
-              consequat. Duis aute irure dolor in reprehenderit in voluptate
-              velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint
-              occaecat cupidatat non proident, sunt in culpa qui officia
-              deserunt mollit anim id est laborum.
-            </li>
-            <br />
-            <li>
-              <strong>Consectetur adipiscing elit</strong> <br /> sed do eiusmod
-              tempor incididunt ut labore et dolore magna aliqua. Ut enim ad
-              minim veniam, quis nostrud exercitation ullamco laboris nisi ut
-              aliquip ex ea commodo consequat. Duis aute irure dolor in
-              reprehenderit in voluptate velit esse cillum dolore eu fugiat
-              nulla pariatur. Excepteur sint occaecat cupidatat non proident,
-              sunt in culpa qui officia deserunt mollit anim id est laborum.
-            </li>
-            <br />
-            <li>
-              <strong>
-                Sed ut perspiciatis unde omnis iste natus error sit voluptatem
-              </strong>
-              <br />
-              accusantium doloremque laudantium, totam rem aperiam, eaque ipsa
-              quae ab illo inventore veritatis et quasi architecto beatae vitae
-              dicta sunt explicabo. Nemo enim ipsam voluptatem quia voluptas sit
-              aspernatur aut odit aut fugit, sed quia consequuntur magni dolores
-              eos qui ratione voluptatem sequi nesciunt. Neque porro quisquam
-              est, qui dolorem ipsum quia dolor sit amet, consectetur, adipisci
-              velit, sed quia non numquam eius modi tempora incidunt ut labore
-              et dolore magnam aliquam quaerat voluptatem. Ut enim ad minima
-              veniam, quis nostrum exercitationem ullam corporis suscipit
-              laboriosam, nisi ut aliquid ex ea commodi consequatur? Quis autem
-              vel eum iure reprehenderit qui in ea voluptate velit esse quam
-              nihil molestiae consequatur, vel illum qui dolorem eum fugiat quo
-              voluptas nulla pariatur?
-            </li>
-            <br />
-            <li>
-              <strong>At vero eos et accusamus</strong>
-              <br />
-              et iusto odio dignissimos ducimus qui blanditiis praesentium
-              voluptatum deleniti atque corrupti quos dolores et quas molestias
-              excepturi sint occaecati cupiditate non provident, similique sunt
-              in culpa qui officia deserunt mollitia animi, id est laborum et
-              dolorum fuga. Et harum quidem rerum facilis est et expedita
-              distinctio. Nam libero tempore, cum soluta nobis est eligendi
-              optio cumque nihil impedit quo minus id quod maxime placeat facere
-              possimus, omnis voluptas assumenda est, omnis dolor repellendus.
-              Temporibus autem quibusdam et aut officiis debitis aut rerum
-              necessitatibus saepe eveniet ut et voluptates repudiandae sint et
-              molestiae non recusandae. Itaque earum rerum hic tenetur a
-              sapiente delectus, ut aut reiciendis voluptatibus maiores alias
-              consequatur aut perferendis doloribus asperiores repellat.
-            </li>
-          </ol>
-        </div>
-      </div>
-    </section>
-  );
+
+
+       return (
+         <section class="section has-text-centered">
+           <SectionTitle name="Terms and Conditions" />
+           <div class="container box is-size-5-mobile">
+             <div className={styles.terms}>
+               <ol>
+                 {termsAndConditions.map((item) => (
+                   <li>
+                     <strong>{item.title}</strong>
+                     <br />
+                     {item.content}
+                     <br />
+                     <br />
+                   </li>
+                 ))}
+               </ol>
+             </div>
+           </div>
+         </section>
+       );
 }