+++ Changelog.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 10/08/2022
+### Added
+ - new workwith us pages (pharmacy, online consultation and laboratory services)
+
+
+10/08/2022
 ### Update
 - New service page
 - New Forms page (validation)